+++ src/map/chat.h
@@ -32,7 +32,7 @@ void chat_changechatstatus(struct map_session_data *sd, unsigned short limit, un
 void chat_kickchat(struct map_session_data *sd, char *kickusername);
 
 int chat_createnpcchat(
-	struct npc_data *nd,int limit,int pub,int trigger,char* title,int titlelen,const char *ev,
+	struct npc_data *nd,int limit,int pub,int trigger,char* title,size_t titlelen,const char *ev,
 	int zeny,int lowlv,int highlv,unsigned int job,int upper);
 int chat_deletenpcchat(struct npc_data *nd);
 int chat_enableevent(struct chat_data *cd);
