+++ src/map/intif.h
@@ -24,8 +24,8 @@
 
 int intif_parse(int fd);
 
-void intif_GMmessage(char* mes, int len, int flag);
-int intif_announce(char* mes,int len,unsigned long color);
+void intif_GMmessage(char* mes, size_t len, int flag);
+int intif_announce(char* mes,size_t len,unsigned long color);
 
 void intif_wis_message(struct map_session_data *sd, char *nick, char *mes, int mes_len);
